Asteroids look okay
