+++ src/CE/MultiMeshManager.cpp
@@ -87,7 +87,10 @@ MultiMeshManager::MultiMeshManager():
   mesh2path.reserve(max_meshes);
 }
 
-MultiMeshManager::~MultiMeshManager() {}
+MultiMeshManager::~MultiMeshManager() {
+  for(auto &it : v_meshes)
+    unused_multimesh(it.second,true);
+}
 
 void MultiMeshManager::time_passed(real_t delta) {
   v_frame++;
@@ -128,7 +131,9 @@ void MultiMeshManager::send_meshes_to_visual_server(real_t projectile_scale,RID 
     if(loud)
       Godot::print("Got a mesh.");
     MeshInfo &mesh_info = vit.second;
-    int count = instance_locations.count(vit.first) + instance_effects.count(vit.first);
+    int lcount=instance_locations.count(vit.first);
+    int icount=instance_effects.count(vit.first);
+    int count = icount+lcount;
     
     if(!count) {
       if(loud) {
@@ -137,10 +142,10 @@ void MultiMeshManager::send_meshes_to_visual_server(real_t projectile_scale,RID 
         else
           Godot::print("Unused mesh found with resource path \""+mesh_info.resource_path+"\"");
       }
-      unused_multimesh(mesh_info);
+      unused_multimesh(mesh_info,false);
       continue;
     }
-
+    
     pair<instlocs_iterator,instlocs_iterator> instlocs =
       instance_locations.equal_range(vit.first);
     pair<insteff_iterator,insteff_iterator> insteffs =
@@ -163,6 +168,15 @@ void MultiMeshManager::send_meshes_to_visual_server(real_t projectile_scale,RID 
     }
     
     pack_visuals(instlocs,insteffs,mesh_info.floats,mesh_info,projectile_scale);
+
+    auto titer = requested_transforms.find(vit.first);
+    if(titer!=requested_transforms.end()) {
+      Transform t = titer->second;
+      requested_transforms.erase(titer);
+      mesh_info.visual_instance_transform = t;
+      visual_server->instance_set_transform(mesh_info.visual_rid,t);
+      Godot::print("Set transform for mesh "+str(vit.first)+" as "+str(t));
+    }
     
     // Send the instance data.
     visual_server->multimesh_set_visible_instances(mesh_info.multimesh_rid,count);
@@ -246,6 +260,7 @@ bool MultiMeshManager::update_visual_instance(MeshInfo &mesh_info,RID scenario,b
     visual_server->instance_set_layer_mask(mesh_info.visual_rid,EFFECTS_LIGHT_LAYER_MASK);
     visual_server->instance_set_visible(mesh_info.visual_rid,true);
     visual_server->instance_geometry_set_cast_shadows_setting(mesh_info.visual_rid,0);
+    visual_server->instance_set_transform(mesh_info.visual_rid,mesh_info.visual_instance_transform);
   } else if(reset_scenario)
     visual_server->instance_set_scenario(mesh_info.visual_rid,scenario);
   return true;
@@ -279,16 +294,16 @@ bool MultiMeshManager::load_mesh(MeshInfo &mesh_info) {
 void MultiMeshManager::clear_all_multimeshes() {
   FAST_PROFILING_FUNCTION;
   for(auto &it : v_meshes)
-    unused_multimesh(it.second);
+    unused_multimesh(it.second,true);
 }
 
-void MultiMeshManager::unused_multimesh(MeshInfo &mesh_info) {
+void MultiMeshManager::unused_multimesh(MeshInfo &mesh_info,bool force) {
   FAST_PROFILING_FUNCTION;
   // No instances in this multimesh. Should we delete it?
   if(!mesh_info.visual_rid.is_valid() and !mesh_info.multimesh_rid.is_valid())
     return;
   
-  if(v_frame > mesh_info.last_frame_used+1200) {
+  if(force or v_frame > mesh_info.last_frame_used+1200) {
     //Godot::print("Freeing a multimesh");
     if(mesh_info.visual_rid.is_valid())
       visual_server->free_rid(mesh_info.visual_rid);
@@ -359,20 +374,21 @@ void MultiMeshManager::pack_visuals(const pair<instlocs_iterator,instlocs_iterat
       p_instance!=effects.second && i<stop;  p_instance++, i+=nfloat) {
     InstanceEffect &instance = p_instance->second;
     const Basis &b = instance.transform.basis;
+    //Basis b;
     const Vector3 &o = instance.transform.origin;
     const Color &c = instance.color_data;
     const Color &d = instance.instance_data;
     dataptr[i + 0] = b[0][0];
     dataptr[i + 1] = b[0][1];
     dataptr[i + 2] = b[0][2];
-    dataptr[i + 3] = b[1][0];
-    dataptr[i + 4] = b[1][1];
-    dataptr[i + 5] = b[1][2];
-    dataptr[i + 6] = b[2][0];
-    dataptr[i + 7] = b[2][1];
-    dataptr[i + 8] = b[2][2];
-    dataptr[i + 9] = o.x;
-    dataptr[i + 10] = o.y;
+    dataptr[i + 3] = o.x;
+    dataptr[i + 4] = b[1][0];
+    dataptr[i + 5] = b[1][1];
+    dataptr[i + 6] = b[1][2];
+    dataptr[i + 7] = o.y;
+    dataptr[i + 8] = b[2][0];
+    dataptr[i + 9] = b[2][1];
+    dataptr[i + 10] = b[2][2];
     dataptr[i + 11] = o.z;
     dataptr[i + 12] = c.r;
     dataptr[i + 13] = c.g;
