+++ src/include/ScriptUtils.hpp
@@ -5,8 +5,15 @@
 #include <Reference.hpp>
 #include <String.hpp>
 #include <Array.hpp>
+#include <ArrayMesh.hpp>
+#include <Image.hpp>
 
 namespace godot {
+
+  Ref<ArrayMesh> make_icosphere(int subs);
+  Ref<ArrayMesh> make_cube_sphere_v2(float radius, int subs);
+  Ref<Image> make_lookup_tiles_c224();
+  Ref<Image> make_lookup_tiles_c112();
 
   class ScriptUtils: public Reference {
     GODOT_CLASS(ScriptUtils,Reference)
@@ -20,6 +27,11 @@ namespace godot {
     void noop() const;
     String string_join(Array a,String sep) const;
     String string_join_no_sep(Array a) const;
+
+    Ref<ArrayMesh> make_icosphere(float radius, int subs) const;
+    Ref<ArrayMesh> make_cube_sphere_v2(float radius, int subs) const;
+    Ref<Image> make_lookup_tiles_c224() const;
+    Ref<Image> make_lookup_tiles_c112() const;
   };
 
 }
