+++ src/include/SphereTool.hpp
@@ -7,20 +7,18 @@
 #include <Ref.hpp>
 
 namespace godot {
-
-class SphereTool: public MeshInstance {
+  
+  class SphereTool: public MeshInstance {
     GODOT_CLASS(SphereTool, MeshInstance)
-
-public:
+    
+  public:
     static void _register_methods();
     SphereTool();
     ~SphereTool();
     void _init();
     void make_icosphere(String name,Vector3 center, float radius, int subs);
     void make_cube_sphere_v2(String name,Vector3 center, float radius, int subs);
-    Ref<Image> make_lookup_tiles_c224() const;
-    Ref<Image> make_lookup_tiles_c112() const;
-};
+  };
 
 }
 #endif
